+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # 1.2.0 - Unknown
 - Fixed USACO parser
-- Added USACO Training parser
-- Removed jQuery dependency
 - Reverted CodeForces parser back to the CHelper one
 - Switched the new Google Code Jam parser to the CHelper one
+- Added parsers for USACO Training, DevSkill, DMOJ, URI Online Judge and LightOJ
 
 # 1.1.0 - April 1st, 2018
 - Added support for the new Google Code Jam website